First cut at a consolidated test_cql.py
